requirements.txt
absl-py==0.10.0
appdirs==1.4.4
astunparse==1.6.3
backcall==0.2.0
bleach==3.2.1
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorama==0.4.4
contextlib2==0.6.0.post1
cycler==0.10.0
decorator==4.4.2
distlib==0.3.1
docutils==0.16
filelock==3.0.12
Flask==1.1.2
gast==0.3.3
google-auth==1.21.0
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.31.0
h5py==2.10.0
idna==2.10
inflection==0.5.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.11.3
Keras==2.4.3
Keras-Preprocessing==1.1.2
keyring==21.5.0
kiwisolver==1.3.1
lxml==4.6.1
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.3.3
more-itertools==8.4.0
numpy==1.18.5
oauthlib==3.1.0
opt-einsum==3.3.0
packaging==20.4
pandas==1.0.5
parso==0.7.1
pickleshare==0.7.5
Pillow==8.0.1
pkginfo==1.6.1
prompt-toolkit==3.0.8
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.7.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
pywin32-ctypes==0.2.0
PyYAML==5.3.1
Quandl==3.5.2
readme-renderer==28.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rsa==4.6
scipy==1.4.1
six==1.15.0
tensorboard==2.3.0
tensorboard-plugin-wit==1.7.0
tensorflow-estimator==2.3.0
termcolor==1.1.0
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
twine==3.2.0
urllib3==1.25.10
virtualenv==20.0.31
virtualenvwrapper-win==1.2.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
wrapt==1.12.1